% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data75}
\alias{data75}
\title{Simulated Exposure data -75\% below LOD}
\format{A data frame with 259 rows and 2 variables - exposure and status

The variables are as follows:
\itemize{
  \item exposure Exposure measurements
  \item Status 0=detect  1= below limit of detection. Exposure values for status=1 is  the limit of detection for that measurement.
}}
\description{
A dataset containing simulated exposure for
for 250 subjects with two different limits of detection with 75\% of the data below LOD
}

