\name{lod-package}
\alias{lod-package}
\alias{lod}
\docType{package}
\title{
Fit Steps for RNA unwiding traces
}
\description{
Impute exposure values for measurements below the limit of detection assuming a Generalized Gamma Distribution
}
\details{
\tabular{ll}{
Package: \tab lod\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-01\cr
License: \tab GPL-2\cr
}
The main command is lod() which estimates exposure values for data below the limit of detection.
}
\author{
Srikesh Arunajadai

Maintainer: Srikesh Arunajadai<sa2658@columbia.edu>
}
\references{
Arunajadai SG., Rauh VA (2012) Handling covariates subject to limits of detection in regression. Volume 19 issue 3 Pages 369-391.
}
