% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lod}
\alias{lod}
\title{Estimates the exposure value for observation below the limit of detection (LOD) assuming
generalized gamma distribution for the exposure.}
\usage{
lod(data, n.impute = 10, control = list(maxit = 2000))
}
\arguments{
\item{data}{Data with two columns - first column is named exposure and second column is named status.
Status takes on two values 0-detect and 1-non-detect i.e. below the limit of detection. The exposure values
for the non-detects is the limit of detection}

\item{n.impute}{The required number of imputed data sets of the exposure data}

\item{control}{List of control parameters that will be passed on to optim. See optim for details.}
}
\value{
Returns a list with the following elements
\itemize{
\item convergence - Indicator if the optimization algorithm to estimate parameters of the
generalized gamma distribution converged. 0 indicates convergence.
\item parameters - The parameter estimates of the generalized gamma distribution
\item se - The standard error estimates of the parameters of the generalized gamma distribution
\item imputed.data - A matrix of imputed values with number of columns equal to n.impute and number
of rows equal to number of rows of data.
}
}
\description{
Estimates the exposure value for observation below the limit of detection (LOD) assuming
generalized gamma distribution for the exposure.
}
\examples{
\dontrun{

my.lod<-lod(data10)
names(my.lod)

plot(my.lod)
my.lod[my.lod$status==1,]
}
}
\references{
Arunajadai SG., Rauh VA (2012) Handling covariates subject to limits of detection in regression. Volume 19 issue 3 Pages 369-391.
}

