% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.lod}
\alias{plot.lod}
\title{Plots  the obsrved vs therotical quantiles based on the generalized gamma distribution
generalized gamma distribution for the exposure.}
\usage{
plot.lod(obj)
}
\description{
Plots  the obsrved vs therotical quantiles based on the generalized gamma distribution
generalized gamma distribution for the exposure.
}
\examples{
\dontrun{

my.lod<-lod(data10)
names(my.lod)

plot(my.lod)
my.lod[my.lod$status==1,]
}
}
\references{
Arunajadai SG., Rauh VA (2012) Handling covariates subject to limits of detection in regression. Volume 19 issue 3 Pages 369-391.
}

